6th Semester/Image Processing/Final Project/requirements.txt
Django==2.2.1
numpy==1.16.3
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
